Break test fixtures into own file
This helps with code clarity; we can more easily distinguish the reused code which is used to set up test state, and the tests themselves.
