+++ tests/tests.py
@@ -1,10 +1,9 @@
 import copy
 import json
 import os
-import shutil
 import tarfile
 import xml.etree.ElementTree as ET
-from unittest.mock import ANY, MagicMock, Mock, PropertyMock, call, patch
+from unittest.mock import ANY, MagicMock, PropertyMock, call, patch
 
 import boto3
 import pytest
@@ -19,66 +18,13 @@ from caselawclient.factories import JudgmentFactory, PressSummaryFactory
 from caselawclient.models.identifiers.neutral_citation import NeutralCitationNumber
 from caselawclient.models.identifiers.press_summary_ncn import PressSummaryRelatedNCNIdentifier
 from caselawclient.models.utilities.aws import S3PrefixString
+from conftest import s3_message_raw, v2_message, v2_message_raw
+from helpers import create_fake_bulk_file, create_fake_tdr_file
 from notifications_python_client.notifications import NotificationsAPIClient
 
 from ds_caselaw_ingester import exceptions, ingester, lambda_function
 
 rollbar.init(access_token=None, enabled=False)
-TDR_TARBALL_PATH = os.path.join(
-    os.path.dirname(__file__),
-    "../aws_examples/s3/te-editorial-out-int/TDR-2022-DNWR.tar.gz",
-)
-
-BULK_TARBALL_PATH = os.path.join(os.path.dirname(__file__), "../aws_examples/s3/te-editorial-out-int/test3.tar.gz")
-
-
-v2_message_raw = """
-    {
-        "properties": {
-            "messageType":
-                "uk.gov.nationalarchives.tre.messages.judgmentpackage.available.JudgmentPackageAvailable",
-            "timestamp": "2023-05-15T09:14:53.791409Z",
-            "function": "staging-tre-judgment-packer-lambda",
-            "producer": "TRE",
-            "executionId": "cc46e39f-76ef-43c9-a6d7-c6b064c3556a",
-            "parentExecutionId": "d26458ae-19a7-4159-8381-805075163198"
-        },
-        "parameters": {
-            "status": "JUDGMENT_PARSE_NO_ERRORS",
-            "reference": "TDR-2022-DNWR",
-            "originator": "TDR",
-            "bundleFileURI": "http://172.17.0.2:4566/te-editorial-out-int/TDR-2022-DNWR.tar.gz",
-            "metadataFilePath": "/metadata.json",
-            "metadataFileType": "Json"
-        }
-    }
-    """
-
-s3_message = {
-    "Records": [
-        {
-            "eventSource": "aws:s3",
-            "s3": {
-                "bucket": {
-                    "name": "staging-tre-court-document-pack-out",
-                },
-                "object": {
-                    "key": "QX/e31b117f-ff09-49b6-a697-7952c7a67384/BULK-0.tar.gz",
-                },
-            },
-        },
-    ],
-}
-v2_message = json.loads(v2_message_raw)
-s3_message_raw = json.dumps(s3_message)
-
-
-def create_fake_tdr_file(*args, **kwargs):
-    shutil.copyfile(TDR_TARBALL_PATH, "/tmp/TDR-2022-DNWR.tar.gz")
-
-
-def create_fake_bulk_file(*args, **kwargs):
-    shutil.copyfile(BULK_TARBALL_PATH, "/tmp/BULK-0.tar.gz")
 
 
 def assert_log_sensible(log):
@@ -90,55 +36,6 @@ def assert_log_sensible(log):
     assert "Ingestion complete" in log
     assert "Invalid XML file" not in log
     assert "No XML file found" not in log
-
-
-@pytest.fixture
-@patch(
-    "ds_caselaw_ingester.lambda_function.Ingest.save_tar_file_in_s3",
-    return_value="/tmp/TDR-2022-DNWR.tar.gz",
-)
-@patch("ds_caselaw_ingester.ingester.uuid4")
-def v2_ingest(mock_uuid4, fake_s3):
-    mock_uuid4.return_value = "v2-a1b2-c3d4"
-    create_fake_tdr_file()
-    return ingester.Ingest(
-        message=lambda_function.Message.from_message(v2_message),
-        destination_bucket="bucket",
-        api_client=Mock(),
-        s3_client=Mock(),
-    )
-
-
-@pytest.fixture
-@patch("ds_caselaw_ingester.lambda_function.Ingest.save_tar_file_in_s3", return_value="/tmp/BULK-0.tar.gz")
-@patch("ds_caselaw_ingester.ingester.uuid4")
-def s3_ingest(mock_uuid4, fake_s3):
-    mock_uuid4.return_value = "s3-a1b2-c3d4"
-    create_fake_bulk_file()
-    return ingester.Ingest(
-        message=lambda_function.Message.from_message(s3_message),
-        destination_bucket="bucket",
-        api_client=Mock(),
-        s3_client=Mock(),
-    )
-
-
-@pytest.fixture
-@patch(
-    "ds_caselaw_ingester.lambda_function.Ingest.save_tar_file_in_s3",
-    return_value="/tmp/TDR-2022-DNWR.tar.gz",
-)
-def fcl_ingest(fake_s3):
-    "Fake a FCL reparse message (badly)"
-    new_message = copy.deepcopy(v2_message)
-    new_message["parameters"]["originator"] = "FCL"
-
-    return ingester.Ingest(
-        message=lambda_function.Message.from_message(new_message),
-        destination_bucket="bucket",
-        api_client=Mock(),
-        s3_client=Mock(),
-    )
 
 
 class TestHandler:
@@ -825,8 +722,8 @@ class TestEmailLogic:
         bulk.assert_not_called()
 
     def test_fcl_ingest_no_email(self, bulk, new, updated, fcl_ingest):
-        v2_ingest.inserted = True
-        v2_ingest.updated = False
+        fcl_ingest.inserted = True
+        fcl_ingest.updated = False
         fcl_ingest.send_email()
 
         updated.assert_not_called()
